requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psutil==5.9.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.0
pydantic-settings==2.1.0
tensorflow==2.15.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
reportlab==4.0.7
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
email-validator==2.1.0
pywin32==306; sys_platform=="win32"
wmi==1.5.1; sys_platform=="win32"
pyttsx3==2.90
speechrecognition==3.10.0
